Fix: no longer require SRC_GITHUB_TOKEN to be set
